requirements-dev.txt
#########################################################
# PAYROLL ANALYZER - ENVIRONNEMENT DE DÉVELOPPEMENT
#########################################################

psycopg[binary,pool]>=3.2
python-dotenv>=1.0.1
pydantic>=2.6
alembic>=1.13
sqlalchemy>=2.0
requests>=2.32
colorama>=0.4.6


# ---- UI / PYQT ----

PyQt6>=6.7
PyQt6-WebEngine>=6.7
qt-material>=2.14
qasync>=0.27


# ---- GRAPHIQUES HTML (lib Python, pas de modif HTML) ----

plotly>=5.24


# ---- TESTS / QUALITÉ ----

pytest>=8.2
pytest-cov>=5.0
faker>=25.0
bandit>=1.7
safety>=3.2
ruff>=0.6
black>=24.8
isort>=5.13
mypy>=1.11
pre-commit>=3.8


# ---- OUTILS CI / AUTOMATION ----

tox>=4.15
commitizen>=3.25


# ---- DÉPENDANCES INTERNES ----

-e .
